cache

+++ python/tools/prepare_build_environment_linux.sh
@@ -1,12 +1,8 @@
 #! /bin/bash
-
 set -e
 set -x
-
 pip install "cmake==3.22.*"
-
 if [ "$CIBW_ARCHS" == "aarch64" ]; then
-
     OPENBLAS_VERSION=0.3.26
     curl -L -O https://github.com/xianyi/OpenBLAS/releases/download/v${OPENBLAS_VERSION}/OpenBLAS-${OPENBLAS_VERSION}.tar.gz
     tar xf *.tar.gz && rm *.tar.gz
@@ -17,7 +13,6 @@ if [ "$CIBW_ARCHS" == "aarch64" ]; then
     make install NO_SHARED=1
     cd ..
     rm -r OpenBLAS-*
-
 else
     dnf install -y dnf-plugins-core
     # Install CUDA 12.8:
@@ -35,42 +30,53 @@ else
         libnccl-2.26.2-1+cuda12.8 \
         libnccl-devel-2.26.2-1+cuda12.8
     ln -s cuda-12.8 /usr/local/cuda
-
     ONEAPI_VERSION=2025.3.0
     dnf config-manager --add-repo https://yum.repos.intel.com/oneapi
     rpm --import https://yum.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB
     dnf install -y intel-oneapi-mkl-devel-$ONEAPI_VERSION
 
-    ONEDNN_VERSION=3.1.1
-    curl -L -O https://github.com/oneapi-src/oneDNN/archive/refs/tags/v${ONEDNN_VERSION}.tar.gz
-    tar xf *.tar.gz && rm *.tar.gz
-    cd oneDNN-*
-    cmake -DCMAKE_BUILD_TYPE=Release -DONEDNN_LIBRARY_TYPE=STATIC -DONEDNN_BUILD_EXAMPLES=OFF -DONEDNN_BUILD_TESTS=OFF -DONEDNN_ENABLE_WORKLOAD=INFERENCE -DONEDNN_ENABLE_PRIMITIVE="CONVOLUTION;REORDER" -DONEDNN_BUILD_GRAPH=OFF .
-    make -j$(nproc) install
-    cd ..
-    rm -r oneDNN-*
+    CACHE_DIR="/project/.cibw-cache"
+    mkdir -p "$CACHE_DIR"
 
+    # oneDNN
+    ONEDNN_VERSION=3.1.1
+    if [ -f "$CACHE_DIR/onednn-${ONEDNN_VERSION}.tar.gz" ]; then
+        tar xf "$CACHE_DIR/onednn-${ONEDNN_VERSION}.tar.gz" -C /usr/local
+    else
+        curl -L -O https://github.com/oneapi-src/oneDNN/archive/refs/tags/v${ONEDNN_VERSION}.tar.gz
+        tar xf *.tar.gz && rm *.tar.gz
+        cd oneDNN-*
+        cmake -DCMAKE_BUILD_TYPE=Release -DONEDNN_LIBRARY_TYPE=STATIC -DONEDNN_BUILD_EXAMPLES=OFF -DONEDNN_BUILD_TESTS=OFF -DONEDNN_ENABLE_WORKLOAD=INFERENCE -DONEDNN_ENABLE_PRIMITIVE="CONVOLUTION;REORDER" -DONEDNN_BUILD_GRAPH=OFF .
+        make -j$(nproc) install
+        cd ..
+        rm -r oneDNN-*
+        tar czf "$CACHE_DIR/onednn-${ONEDNN_VERSION}.tar.gz" -C /usr/local lib/libdnnl* lib/cmake/dnnl include/dnnl* include/oneapi
+    fi
+
+    # OpenMPI
     OPENMPI_VERSION=4.1.6
-    curl -L -O https://download.open-mpi.org/release/open-mpi/v4.1/openmpi-${OPENMPI_VERSION}.tar.bz2
-    tar xf *.tar.bz2 && rm *.tar.bz2
-    cd openmpi-*
-    ./configure
-    make -j$(nproc) install
-    cd ..
-    rm -r openmpi-*
+    if [ -f "$CACHE_DIR/openmpi-${OPENMPI_VERSION}.tar.gz" ]; then
+        tar xf "$CACHE_DIR/openmpi-${OPENMPI_VERSION}.tar.gz" -C /usr/local
+    else
+        curl -L -O https://download.open-mpi.org/release/open-mpi/v4.1/openmpi-${OPENMPI_VERSION}.tar.bz2
+        tar xf *.tar.bz2 && rm *.tar.bz2
+        cd openmpi-*
+        ./configure
+        make -j$(nproc) install
+        cd ..
+        rm -r openmpi-*
+        tar czf "$CACHE_DIR/openmpi-${OPENMPI_VERSION}.tar.gz" -C /usr/local lib/libmpi* lib/libopen* lib/openmpi lib/pkgconfig/ompi* include/mpi* include/openmpi bin/mpi* bin/opal* bin/ompi* bin/orte* share/openmpi
+    fi
+
     export LD_LIBRARY_PATH="/usr/local/lib/:$LD_LIBRARY_PATH"
 fi
-
 mkdir build-release && cd build-release
-
 if [ "$CIBW_ARCHS" == "aarch64" ]; then
     cmake -DCMAKE_BUILD_TYPE=Release -DBUILD_CLI=OFF -DWITH_MKL=OFF -DOPENMP_RUNTIME=COMP -DCMAKE_PREFIX_PATH="/opt/OpenBLAS" -DWITH_OPENBLAS=ON -DWITH_RUY=ON ..
 else
     cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_CXX_FLAGS="-msse4.1" -DBUILD_CLI=OFF -DWITH_DNNL=ON -DOPENMP_RUNTIME=COMP -DWITH_CUDA=ON -DWITH_CUDNN=OFF -DCUDA_DYNAMIC_LOADING=ON -DCUDA_NVCC_FLAGS="-Xfatbin=-compress-all" -DCUDA_ARCH_LIST="Common"  -DWITH_TENSOR_PARALLEL=ON ..
 fi
-
 VERBOSE=1 make -j$(nproc) install
 cd ..
 rm -r build-release
-
 cp README.md python/
